Refactored memory field of view handler to be abstract and allow more generic effects.  Added a handler for dimming colors based on a multiplier when they are out of FOV.
